backend_streamlit/requirements.txt
requests
<<<<<<< HEAD
streamlit
python-dotenv
=======
streamlit
>>>>>>> af586de9ea55b2a5ddbbf0e0a05a8632e88d8386
